fix: 修复 git pull 失败仍然重命名远程分支问题 git pull 失败原因：代码冲突，当发生pull失败，skip后面的操作

+++ git-tool.sh
@@ -17,13 +17,13 @@ function rename_remote()
 		fi
 	fi
 
-	git pull
-	git push --delete origin $oldBranchName
-	git branch -m $oldBranchName $newBranchName
-	git push origin $newBranchName
-	git branch --set-upstream-to=origin/$newBranchName $newBranchName
-
-	echo "done"
+	if git pull ; then
+		git push --delete origin $oldBranchName
+		git branch -m $oldBranchName $newBranchName
+		git push origin $newBranchName
+		git branch --set-upstream-to=origin/$newBranchName $newBranchName
+		echo "done"
+	fi	
 }
 
 function usage()
